+++ src/mailman/rules/tests/test_moderation.py
@@ -26,7 +26,7 @@ from mailman.interfaces.member import MemberRole
 from mailman.interfaces.usermanager import IUserManager
 from mailman.rules import moderation
 from mailman.testing.helpers import (
-    set_preferred, specialized_message_from_string as mfs)
+    LogFileMark, set_preferred, specialized_message_from_string as mfs)
 from mailman.testing.layers import ConfigLayer
 from zope.component import getUtility
 
@@ -106,6 +106,34 @@ A message body.
         reasons = msgdata['moderation_reasons']
         self.assertEqual(
             reasons, ['The message comes from a moderated member'])
+
+    def test_these_nonmembers_malformed_regexp(self):
+        mark = LogFileMark('mailman.error')
+        user_manager = getUtility(IUserManager)
+        nonmembers = [
+            # Malformed regexps should be skipped
+            '^*@broken.regex',
+            'anne@example.com',
+            ]
+        rule = moderation.NonmemberModeration()
+        user_manager = getUtility(IUserManager)
+        user_manager.create_address('anne@example.com')
+        setattr(self._mlist, 'accept_these_nonmembers', nonmembers)
+        msg = mfs("""\
+From: anne@example.com
+To: test@example.com
+Subject: A test message
+Message-ID: <ant>
+MIME-Version: 1.0
+
+A message body.
+""")
+        msgdata = {}
+        result = rule.check(self._mlist, msg, msgdata)
+        self.assertFalse(result, 'NonmemberModeration rule should not hit')
+        self.assertIn("Invalid regexp '^*@broken.regex' in "
+                      'accept_these_nonmembers for test.example.com: '
+                      'nothing to repeat', mark.readline())
 
     def test_these_nonmembers(self):
         # Test the legacy *_these_nonmembers attributes.